+++ src/phasor/Rule.ts
